tests: pull in hghave
Previously, the hghave checks that were commented out in the tests were broken
if uncommented.  One cause was that it was expecting hghave in the testdir,
while our testdir didn't contain hghave.  Now it does.

The hghave was pulled unmodified from Mercurial 2.3, to match the version of
run-tests.py in use.