tests: avoid changing the current directory
Mercurial allows specifying which repository to use via the -R/--repository
option.  Git allows a similar function using the --git-dir option.  By using
these options, in many cases we can avoid checking the current directory.
This makes tests easier to understand, as you don't need to remember which
directory you're in to understand what's going on.  It also makes tests easier
to write, as you don't need to remember to cd out of a directory when you're
done doing things there.

Thanks to Felipe Contreras for the patch which this was based on.