push: provide better output about changed references (issue #64)
As pointed out by l33t, Hg-Git's output for push doesn't currently do a very
good job of telling the user what happened.  My previous changes in this area
had moved some of the output from status to note, making it only show if
--verbose was specified.  However, I hadn't realized at the time that the
reference information (though overly verbose) was providing a valueable purpose
that otherwise wasn't met; telling the user that a remote reference had changed.

This changeset makes it so that:
*   default output will include simple messages like "adding reference
    refs/heads/feature" and "updating reference refs/heads/master" (omitting any
    mention of unchanged references)
*   verbose output will include more detailed messages like "adding reference
    default::refs/heads/feature => GIT:aba43c" and "updating reference
    default::refs/heads/master => GIT:aba43c" (omitting any mention of unchanged
    references)
*   debug output will include the detailed output like in verbose, but
    addtionally will include messages like "unchanged reference
    default::refs/heads/other => GIT:aba43c"

https://bitbucket.org/durin42/hg-git/issue/64/push-confirmation