listkeys: while looking for bookmarks, don't fetch a pack
hg pull calls listkeys for bookmarks. This would previously cause a pack with
all refs to be fetched. For Mercurial mirrors of Git repositories where only
some refs were mirrored, this would cause problems in a bunch of ways:

- A larger pack would be fetched than necessary.
- The final refs written out to the Git repo would only be the set of refs we
  were actually interested in. If a GC was subsequently run, unreferenced
  objects would be deleted. Those objects might be referred to on subsequent
  fetches, which could cause hg-git to crash.

We replace all that logic with a simple null fetch. The tests introduced in the
previous patch ensure no regressions.