auth: use a ~/.git-credentials file if it exists to fetch credentials
In some situations we code the plain auth-less uri in the hgrc paths
section (for convenience) and want to use a centrally-stored
credentials file.

Though it is possible to use the hgrc [auth] section, it would be also
nice to access .git-credentials (still usefull for git direct access),
avoiding double maintenance of credentials.

This probably does not cover all possible situations with git
credentials handling but it provides a working base.

closes #305