heptapod-ci: protecting outer commands from PYTHONPATH
All commands that make use of Python and are meant to retrieve or prebuild
code should not themselves use potentially incompatible stuff from
/ci/repos/mercurial.

For instance, updating Mercurial to 'stable' should not risk using a
random version of the mercurial Python package. It must instead use
the one provided system-wide, that is consistent with the `hg` command.