py3: avoid using repr() on bytes
We didn?t realize that this was broken because tests/test-illegal-contents.t is
skipped on the CI.

The change is similar to 503e9abfd50e.

As can be seen in the doctest, this changes output of non-ASCII characters. I
argue that the new output is better than the old because the old output exposed
an implementation detail (the repr() function of Python) while the new output
is something that can be copied and passed to other programs.