hgrepo: don't require wlock for writing git metadata
The Git metadata files are located directly within '.hg' in the store
repository. To use the vfs directly, we have to lock the working
directory. The Mercurial push and pull commands ? reasonably ? don't
do that, so without this change, we cannot save or load the map during
them.

Ideally, we'd move these files into the store, but that's too
invasive, and would break backward compatibility rather significantly,
unless we add some form of upgrade/downgrade support. Instead, we
simply add them to the list of files that don't need the lock.