tests: don't rely on cloning into an empty directory
Mercurial 5.8 and earlier would delete such a directory on errors,
and the test accidentally relied on that behaviour. To work around
that, we clone into the _current_ directory, which Mercurial never
deletes.

The test now reveals the incorrect behaviour, but the current design
makes that difficult to address without tracking unnecessary state.