push: abort with an error when no bookmarks or tags exist
Previously, we would simply invent a master bookmark when this
occurred, but in the specific case where the remote is empty. I
believe that's actually unhelpful in the long run: Should a user not
know about bookmarks, the first push will succeed, and subsequent
pushes then not do anything. How did that bookmark get there?

Instead, we should trigger a hard error for this case; that way, we'll
hopefully cause the user to take a step back and find out how the
command really works.