push: adjust error message when pushing with missing git commits
Thorsten Kampe reported that he ran into this message in issue #376.
We had some handling for this previously, but we would only handle
converting commits with missing parents; this fixes it so that we also
handle pushing truly missing commits rather than only when actually
converting them.

The previous error message was confusing and unhelpful, though, but
this should make it clear what's going on. Fortunately, he _also_
provided a relatively simple way to trigger it, and as it was
previously uncovered by the test suite, I added a test that triggers
it!