versioning: create branch 1.0.x
The upcoming release will be called 1.0 rather than 0.11.0. The main
motivation for this is that I consider hg-git ?ready?. In order for
semantic versioning to make sense, something has to eventually be good
enough. With the changes to user experience, I personally consider
this good enough to actually use. And besides, if this isn't 1.0, when
should it be? When development ceases?

I've updated `CONTRIBUTING.rst` to specify an explicit compatibility
policy.