tests: use an intermediate bare repository
This removes the need to set configure `receive.denyCurrentBranch` in
Git, which is in part bad form, and also explicitly covered
`test-push-to-head.t`. Pushing to a checked out branch in a non-bare
repository is confusing, and leads to weird results in Git.

More than anything, this is just a slight simplification.